+++ apps/nextjs/src/components/features/modals/file-viewer-modal/header.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { ComingSoonTooltip } from '@/components/features/tooltips';
+import { ComingSoonTooltip } from '@/components/features/tooltips/coming-soon-tooltip';
 import { Button } from '@/components/ui/button';
 import { Flex } from '@/components/ui/flex';
 import { Icon } from '@/components/ui/icon';
