+++ apps/nextjs/src/components/features/projects/project-list-item/project-list-item-new.tsx
@@ -1,4 +1,4 @@
-import { ComingSoonTooltip } from '@/components/features/tooltips';
+import { ComingSoonTooltip } from '@/components/features/tooltips/coming-soon-tooltip';
 import { Flex, type FlexProps } from '@/components/ui/flex';
 import { NewBox } from '@/components/ui/new-box';
 import { Text } from '@/components/ui/text';
