--- apps/nextjs/src/components/features/tooltips/index.ts
@@ -1,1 +1,0 @@
-export * from './coming-soon-tooltip';
