+++ apps/nextjs/src/components/pages/workspaces/components/overview/overview-content/key-resources-section/key-resources-empty/key-resources-empty.tsx
@@ -1,4 +1,4 @@
-import { ComingSoonTooltip } from '@/components/features/tooltips';
+import { ComingSoonTooltip } from '@/components/features/tooltips/coming-soon-tooltip';
 import { Button } from '@/components/ui/button';
 import { Flex } from '@/components/ui/flex';
 import { Heading } from '@/components/ui/heading';
